requirements.txt
absl-py==2.3.1
area==1.1.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
babel==2.17.0
backrefs==5.9
beautifulsoup4==4.12.3
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.0
coverage==7.9.2
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fiona==1.10.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gast==0.6.0
geopandas==1.0.1
ghp-import==2.1.0
google-pasta==0.2.0
griffe==1.7.3
grpcio==1.73.1
h5py==3.14.0
hjson==3.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.2
mkdocs-get-deps==0.2.0
mkdocs-macros-plugin==1.3.7
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
mkdocs-ultralytics-plugin==0.1.25
mkdocstrings==0.29.1
mkdocstrings-python==1.16.12
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.2.1
numpy==1.24.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
paginate==0.5.7
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
progressbar2==4.5.0
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycoral @ file:///home/alf/edapa/ultralytics/pycoral-2.0.0-cp39-cp39-linux_x86_64.whl#sha256=77c81c64a99119019c0d65ae9b1af25d2856ab6057dac27d3ea64dac963bef16
Pygments==2.19.2
pymdown-extensions==10.16
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.6.1
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.4
rich==14.0.0
scipy==1.13.1
shapely==2.0.7
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
super_collections==0.5.3
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tflite-runtime @ file:///home/alf/edapa/ultralytics/tflite_runtime-2.5.0.post1-cp39-cp39-linux_x86_64.whl#sha256=132f8c184b99d69e40040d8067eb962d9435cd2579f5563c33b8673d257d353d
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics @ file:///home/alf/edapa/ultralytics
ultralytics-thop==2.0.14
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
